requirements.txt
Django==5.1
djangorestframework==3.15.2
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.24
langchain-openai==0.1.19
langchain-text-splitters==0.2.2
langgraph==0.1.16
pandas==2.2.2
pydantic==2.8.2
python-dotenv==1.0.1
supabase==2.6.0
tabulate==0.9.0
pypdf
